Seeds are non-negative now.

+++ FsCsMvc41/FsCsMvc41WebApp/HomeController.fs
@@ -5,7 +5,7 @@ open System.Web
 open System.Web.Mvc
 
 type IndexModel = { 
-    Seed: int
+    Seed: int64
     MazeHeight: int
     MazeWidth: int
 }
@@ -13,8 +13,8 @@ type IndexModel = {
 [<HandleError>]
 type HomeController() =
     inherit Controller()
-    member this.Index (seed:Nullable<int>, hallWidth:Nullable<int>, rooms:Nullable<int>) =
-        let seed = if seed.HasValue then seed.Value else (int)DateTime.Now.Ticks
+    member this.Index (seed:Nullable<int64>, hallWidth:Nullable<int>, rooms:Nullable<int>) =
+        let seed = if seed.HasValue then seed.Value else DateTime.Now.Ticks
 
         this.View({Seed = seed; MazeHeight = 520; MazeWidth = 520}) :> ActionResult
